+++ modin/distributed/dataframe/pandas/partitions.py
@@ -248,3 +248,39 @@ def from_partitions(
         frame.synchronize_labels(axis=labels_axis_to_sync)
 
     return DataFrame(query_compiler=PandasQueryCompiler(frame))
+
+
+def repartition(df, axis: Optional[int] = None):  # type:ignore
+    """
+    Repartitioning Modin objects to get ideal partitions inside.
+
+    Allows to improve performance where the query compiler cannot yet.
+
+    Parameters
+    ----------
+    df : DataFrame or Series
+    axis : int, optional
+
+    Returns
+    -------
+    DataFrame or Series
+        The repartitioned dataframe or series, depending on the original type.
+    """
+    if axis not in (0, 1, None):
+        raise NotImplementedError
+
+    from modin.pandas import DataFrame, Series
+
+    if not isinstance(df, (DataFrame, Series)):
+        raise NotImplementedError
+
+    list_axis = [0, 1] if axis is None else [axis]
+
+    new_query_compiler = df._query_compiler
+    for _ax in list_axis:
+        new_query_compiler = new_query_compiler.__constructor__(
+            new_query_compiler._modin_frame.apply_full_axis(
+                _ax, lambda df: df, keep_partitioning=False
+            )
+        )
+    return df.__constructor__(query_compiler=new_query_compiler)
