+++ modin/test/test_utils.py
@@ -18,7 +18,6 @@ import pytest
 
 from modin.error_message import ErrorMessage
 import modin.utils
-import modin.pandas as pd
 
 
 # Note: classes below are used for purely testing purposes - they
@@ -289,39 +288,3 @@ def test_warns_that_defaulting_to_pandas():
 
     with warns_that_defaulting_to_pandas():
         ErrorMessage.default_to_pandas(message="Function name")
-
-
-@pytest.mark.parametrize("axis", [0, 1, None])
-@pytest.mark.parametrize("dtype", ["DataFrame", "Series"])
-def test_repartition(axis, dtype):
-    df = pd.DataFrame({"col1": [1, 2], "col2": [5, 6]})
-    df2 = pd.DataFrame({"col3": [9, 4]})
-
-    df = pd.concat([df, df2], axis=1)
-    df = pd.concat([df, df], axis=0)
-
-    obj = df if dtype == "DataFrame" else df["col1"]
-
-    source_shapes = {
-        "DataFrame": (2, 2),
-        "Series": (2, 1),
-    }
-    # check that the test makes sense
-    assert obj._query_compiler._modin_frame._partitions.shape == source_shapes[dtype]
-
-    obj = modin.utils.repartition(obj, axis=axis)
-
-    if dtype == "DataFrame":
-        results = {
-            None: (1, 1),
-            0: (1, 2),
-            1: (2, 1),
-        }
-    else:
-        results = {
-            None: (1, 1),
-            0: (1, 1),
-            1: (2, 1),
-        }
-
-    assert obj._query_compiler._modin_frame._partitions.shape == results[axis]
